refactor rest endpoint generation
